+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export const VERSION = "0.6.21";
+export const VERSION = "0.6.22";
 
 export type {
 	CreateComponentArgs,